refactor: ranking 프래그먼트 삭제 및 switchcase로 변경
