+++ src/pages/ranking/style.ts
@@ -1,0 +1,38 @@
+import styled from "styled-components";
+
+export const Container = styled.div`
+    display: flex;
+    flex-direction: column;
+    gap:24px;
+`
+
+export const TopRankSection = styled.div`
+    width:100vw;
+    max-width:1416px;
+    gap:24px;
+    padding:32px;
+    display:flex;
+    flex-direction:column;
+    align-items:center;
+    box-shadow: 0px 2px 4px -1px #0c0c0d1a, 0px 2px 4px -1px #0c0c0d0d;
+`
+
+export const TopRankContainer = styled.div`
+    width:100%;
+    max-width:1352px;
+    display:flex;
+    align-items:flex-end;
+    justify-content:center;
+    gap:120px;
+`
+
+export const BottomRankSection = styled.div`
+    width:100vw;
+    max-width:1416px;
+    gap:24px;
+    padding:32px;
+    display:flex;
+    flex-direction:column;
+    align-items:center;
+    box-shadow: 0px 2px 4px -1px #0c0c0d1a, 0px 2px 4px -1px #0c0c0d0d;
+`
